version fix

+++ migrations/20241025120845_pre_registered.up.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS pre_registered (
   verification_code_hashed VARCHAR(64) NOT NULL,
   salt VARCHAR(100) NOT NULL,
   was_used BOOLEAN NOT NULL DEFAULT FALSE,
-  version INT NOT NULL DEFAULT 1,
+  version SMALLINT NOT NULL DEFAULT 1,
   created_at TIMESTAMP
   WITH
     TIME ZONE NOT NULL DEFAULT NOW (),
