+++ migrations/20241025121103_user_sessions.up.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS user_sessions (
   viewer_id UUID REFERENCES viewers (id) ON DELETE CASCADE NOT NULL,
   hashed_session_token VARCHAR(100) NOT NULL,
   salt VARCHAR(100) NOT NULL,
-  version INT NOT NULL DEFAULT 1,
+  version SMALLINT NOT NULL DEFAULT 1,
   created_at TIMESTAMP
   WITH
     TIME ZONE NOT NULL DEFAULT NOW (),
