+++ migrations/20241025121350_reset_password.up.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS reset_password (
   hashed_reset_password_token VARCHAR(100) NOT NULL,
   salt VARCHAR(100) NOT NULL,
   was_used BOOLEAN NOT NULL DEFAULT FALSE,
-  version INT NOT NULL DEFAULT 1,
+  version SMALLINT NOT NULL DEFAULT 1,
   created_at TIMESTAMP
   WITH
     TIME ZONE NOT NULL DEFAULT NOW (),
